requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pathlib2>=2.3.7
typing-extensions>=4.5.0

# Document processing
python-docx>=0.8.11
beautifulsoup4>=4.12.0

# ML/AI - Core libraries
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Knowledge graph
networkx>=3.1.0

# LLM APIs
google-generativeai>=0.3.0
openai>=1.3.0
anthropic>=0.7.0

# Web UI
gradio>=3.50.0

# Data processing
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.65.0

# Utilities and formatting
rich>=13.5.0
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-mock>=3.11.0

# Optional but recommended
matplotlib>=3.7.0
seaborn>=0.12.0

# Development tools
ipython>=8.12.0
jupyter>=1.0.0